+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the "tsschema" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## 1.0.1
+
+- Allow generating schema for a type without path by using `tsconfig.json`
+- Allow passing parameters to schema generator by url query
+
 ## 1.0.0
 
 - Initial release with `tsschema://` virtual JSON schema capability. Allows file relative and workspace relative paths.